backend/requirements.txt
FastAPI
uvicorn
python-multipart
pydantic
Pillow
numpy
